adding md5 checksum

+++ scripts/download-data.sh
@@ -13,7 +13,27 @@ mkdir -p $resultsdir
     read
     while IFS=, read -r url filename date time size
     do
-        echo "Downloading $filename from $url"
-        curl -L $url/$filename -o $datadir/$filename
+        if [ ! -f $datadir/$filename ]; then
+            echo "Downloading $filename from $url"
+            curl -L $url/$filename -o $datadir/$filename
+
+            echo "Downloading $filename.md5 from $url"
+            curl -L $url/${filename}.md5 -o $datadir/${filename}.md5
+        fi
     done
 } < files.csv
+
+# Navigate to the data directory
+workdir=$(pwd)
+cd $datadir
+
+# Check md5sums
+echo "" > ${resultsdir}/md5sums.txt
+{
+    read
+    while IFS=, read -r url filename date time size
+    do
+        echo "Verifying md5sum for $filename"
+        md5sum -c ${filename}.md5 >> ${resultsdir}/md5sums.txt
+    done
+} < $workdir/files.csv
